split out into modules, introduce Ident newtype
