+++ Cargo.toml
@@ -11,3 +11,4 @@ pest = "2.7.5"
 pest_derive = "2.7.5"
 lazy_static = "1.4.0"
 petgraph = "0.6.4"
+err-derive = "0.3.1"
