+++ src/ast/mod.rs
@@ -1,0 +1,8 @@
+pub mod declaration;
+pub mod error;
+pub mod expression;
+pub mod program;
+
+pub use declaration::Declaration;
+pub use expression::{Expr, Ident, Opcode, UOpcode};
+pub use program::Program;
