+++ src/interpreter.rs
@@ -1,8 +1,8 @@
 use std::collections::HashMap;
 
-use crate::ast::{Declaration, Expr, Opcode, Program, UOpcode};
+use crate::ast::{Declaration, Expr, Ident, Opcode, Program, UOpcode};
 
-pub fn interpret_expr(context: &HashMap<String, Expr>, expr: &Expr) -> i32 {
+pub fn interpret_expr(context: &HashMap<Ident, Expr>, expr: &Expr) -> i32 {
     match expr {
         Expr::Number(n) => *n,
         Expr::UnaryOp(op, expr) => {
@@ -23,16 +23,16 @@ pub fn interpret_expr(context: &HashMap<String, Expr>, expr: &Expr) -> i32 {
         }
         Expr::Variable(name) => match context.get(name) {
             Some(expr) => interpret_expr(context, expr),
-            None => panic!("Variable {} not found in context", name),
+            None => panic!("Variable {:?} not found in context", name),
         },
     }
 }
 
-pub fn interpret(initial_context: HashMap<String, i32>, program: Program) -> i32 {
+pub fn interpret(initial_context: HashMap<Ident, i32>, program: Program) -> i32 {
     let mut context = initial_context
         .iter()
         .map(|(k, v)| (k.clone(), Expr::Number(*v)))
-        .collect::<HashMap<String, Expr>>();
+        .collect::<HashMap<Ident, Expr>>();
     for decl in &program.decls {
         match decl {
             Declaration::VarAssignment(name, expr) => {
@@ -40,7 +40,7 @@ pub fn interpret(initial_context: HashMap<String, i32>, program: Program) -> i32
             }
             Declaration::PublicVar(name) => {
                 if !context.contains_key(name) {
-                    panic!("hey! you forgot to give me a value for {name}")
+                    panic!("hey! you forgot to give me a value for {:?}", name)
                 }
             }
         }
@@ -52,21 +52,12 @@ pub fn interpret(initial_context: HashMap<String, i32>, program: Program) -> i32
 #[cfg(test)]
 mod interpreter_tests {
     use super::*;
-    use crate::ast::Expr;
     use crate::parser;
-    use pest::error::Error;
-    use pest::Parser;
-
-    pub fn parse_single_expression(input: &str) -> Result<Expr, Error<parser::Rule>> {
-        let mut pairs = parser::CalcParser::parse(parser::Rule::expression, input)?;
-        let pair = pairs.next().unwrap();
-        Ok(parser::parse_expr(pair.into_inner()))
-    }
 
     #[test]
     fn no_parens_test() {
         let input = "22 * 44 + 66";
-        let expr = parse_single_expression(input).unwrap();
+        let expr = parser::parse_single_expression(input).unwrap();
         let mut context = HashMap::new();
         assert_eq!(interpret_expr(&mut context, &expr), 1034);
     }
@@ -74,7 +65,7 @@ mod interpreter_tests {
     #[test]
     fn parens_test() {
         let input = "22 * (44 + 66)";
-        let expr = parse_single_expression(input).unwrap();
+        let expr = parser::parse_single_expression(input).unwrap();
         let context = HashMap::new();
         assert_eq!(interpret_expr(&context, &expr), 2420);
     }
@@ -82,16 +73,8 @@ mod interpreter_tests {
     #[test]
     fn pow_test() {
         let input = "2^4 + 1";
-        let expr = parse_single_expression(input).unwrap();
+        let expr = parser::parse_single_expression(input).unwrap();
         let context = HashMap::new();
         assert_eq!(interpret_expr(&context, &expr), 17);
-    }
-
-    #[test]
-    fn complex_test() {
-        let input = "2^(4 +1 )  *  3+ (  2 + 1)^2";
-        let expr = parse_single_expression(input).unwrap();
-        let context = HashMap::new();
-        assert_eq!(interpret_expr(&context, &expr), 32 * 3 + 9);
     }
 }
