+++ src/parser/mod.rs
@@ -1,0 +1,5 @@
+pub mod parser;
+
+pub use parser::parse;
+pub use parser::parse_single_expression;
+pub use parser::CalcParser;
